requirements.txt
jinja2
pyyaml
pillow
paramiko